+++ examples/stable-diffusion/main.py
@@ -3,6 +3,8 @@ import sys
 import os
 
 from diffusers import StableDiffusionPipeline
+import torch
+from torch import autocast
 
 device = "cuda"
 
@@ -14,6 +16,8 @@ def dummy(images, **kwargs):
 # Read prompt from command line
 prompt = " ".join(sys.argv[1:])
 
+# if your are limited by GPU memory and have less than 10GB of GPU RAM available, you can use fp16 for this example just like this line below
+# pipe = StableDiffusionPipeline.from_pretrained("CompVis/stable-diffusion-v1-4", torch_dtype=torch.float16, revision="fp16",  use_auth_token=os.environ['HUGGINGFACE_TOKEN'])
 pipe = StableDiffusionPipeline.from_pretrained(
     "CompVis/stable-diffusion-v1-4", use_auth_token=os.environ["HUGGINGFACE_TOKEN"]
 )
@@ -23,5 +27,6 @@ pipe.safety_checker = dummy
 # Run until we exit with CTRL+C
 while True:
     n = random.randint(1000, 9999)
-    image = pipe(prompt, guidance_scale=7.5)["sample"][0]
+    with autocast("cuda"):
+        image = pipe(prompt, guidance_scale=7.5).images[0]
     image.save(f"{n}.jpeg")